update-log-component(1) Update Log component and Show page to properly distribute state and useEffect API calls
